// simple-minter/types/types.ts
export enum EInputType {
  TITLE = 'title',
  DESCRIPTION = 'description',
  MINT_AMOUNT = 'mintAmount',
  MAIN_IMAGE = 'mainImage',
  FOREVER_MEDIA = 'foreverMedia',
  FOREVER_DOCUMENT = 'foreverDocument',
  TAGS = 'tags',
  CATEGORY = 'category',
  EXTRA = 'extra',
  LOCATION = 'location',
  WEBSITE = 'website',
  CALENDAR = 'calendar',
  FOREVER_ROYALTIES = 'foreverRoyalties',
  SPLIT_REVENUE = 'splitRevenue',
  CUSTOM_KEY_VALUE = 'customKeyValue',
}